requirements.txt
dash
dash-bootstrap-components
flask
lask_sqlalchemy
flask_bootstrap
flask_assets
fuzzywuzzy
nltk
numpy
pandas
requests
transformers
torch
torchvision
